Make the DNS Timeouts caching configurable
This PR adds a config option to allow disabling the caching of domains,
which timed out during a DNS query. By default the option is enabled to
preserve the current behaviour. Setting the option to 0 disables the
cache and the query will be sent the the DNS resolver each time,
regardless of previous timeouts.